+++ lib/src/theme/button/button_size_properties.dart
@@ -24,7 +24,7 @@ class MoonButtonSizeProperties extends ThemeExtension<MoonButtonSizeProperties> 
     height: MoonSizes.sizes.sm,
     iconSizeValue: MoonSizes.sizes.xs,
     padding: EdgeInsets.symmetric(horizontal: MoonSizes.sizes.x3s),
-    textStyle: MoonTextStyles.heading.text14,
+    textStyle: MoonTextStyles.heading.textDefault,
   );
 
   static final md = MoonButtonSizeProperties(
@@ -33,7 +33,7 @@ class MoonButtonSizeProperties extends ThemeExtension<MoonButtonSizeProperties> 
     height: MoonSizes.sizes.md,
     iconSizeValue: MoonSizes.sizes.xs,
     padding: EdgeInsets.symmetric(horizontal: MoonSizes.sizes.x2s),
-    textStyle: MoonTextStyles.heading.text14,
+    textStyle: MoonTextStyles.heading.textDefault,
   );
 
   static final lg = MoonButtonSizeProperties(