+++ lib/src/theme/checkbox/checkbox_colors.dart
@@ -2,6 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'package:moon_design/src/theme/colors.dart';
+import 'package:moon_design/src/theme/typography/typography.dart';
 import 'package:moon_design/src/utils/color_premul_lerp.dart';
 
 @immutable
@@ -11,6 +12,7 @@ class MoonCheckboxColors extends ThemeExtension<MoonCheckboxColors> with Diagnos
     borderColor: MoonColors.light.trunks,
     checkColor: MoonColors.light.goten,
     inactiveColor: Colors.transparent,
+    textColor: MoonTypography.light.colors.bodyPrimary,
   );
 
   static final dark = MoonCheckboxColors(
@@ -18,6 +20,7 @@ class MoonCheckboxColors extends ThemeExtension<MoonCheckboxColors> with Diagnos
     borderColor: MoonColors.dark.trunks,
     checkColor: MoonColors.dark.goten,
     inactiveColor: Colors.transparent,
+    textColor: MoonTypography.dark.colors.bodyPrimary,
   );
 
   /// Checkbox active color.
@@ -32,11 +35,15 @@ class MoonCheckboxColors extends ThemeExtension<MoonCheckboxColors> with Diagnos
   /// Checkbox inactive color.
   final Color inactiveColor;
 
+  /// Checkbox text color.
+  final Color textColor;
+
   const MoonCheckboxColors({
     required this.activeColor,
     required this.borderColor,
     required this.checkColor,
     required this.inactiveColor,
+    required this.textColor,
   });
 
   @override
@@ -45,12 +52,14 @@ class MoonCheckboxColors extends ThemeExtension<MoonCheckboxColors> with Diagnos
     Color? borderColor,
     Color? checkColor,
     Color? inactiveColor,
+    Color? textColor,
   }) {
     return MoonCheckboxColors(
       activeColor: activeColor ?? this.activeColor,
       borderColor: borderColor ?? this.borderColor,
       checkColor: checkColor ?? this.checkColor,
       inactiveColor: inactiveColor ?? this.inactiveColor,
+      textColor: textColor ?? this.textColor,
     );
   }
 
@@ -63,6 +72,7 @@ class MoonCheckboxColors extends ThemeExtension<MoonCheckboxColors> with Diagnos
       borderColor: colorPremulLerp(borderColor, other.borderColor, t)!,
       checkColor: colorPremulLerp(checkColor, other.checkColor, t)!,
       inactiveColor: colorPremulLerp(inactiveColor, other.inactiveColor, t)!,
+      textColor: colorPremulLerp(textColor, other.textColor, t)!,
     );
   }
 
@@ -74,6 +84,7 @@ class MoonCheckboxColors extends ThemeExtension<MoonCheckboxColors> with Diagnos
       ..add(ColorProperty("activeColor", activeColor))
       ..add(ColorProperty("borderColor", borderColor))
       ..add(ColorProperty("checkColor", checkColor))
-      ..add(ColorProperty("inactiveColor", inactiveColor));
+      ..add(ColorProperty("inactiveColor", inactiveColor))
+      ..add(ColorProperty("textColor", textColor));
   }
 }