+++ lib/src/theme/modal/modal_colors.dart
@@ -2,19 +2,31 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'package:moon_design/src/theme/colors.dart';
+import 'package:moon_design/src/theme/icons/icon_theme.dart';
+import 'package:moon_design/src/theme/typography/typography.dart';
 import 'package:moon_design/src/utils/color_premul_lerp.dart';
 
 @immutable
 class MoonModalColors extends ThemeExtension<MoonModalColors> with DiagnosticableTreeMixin {
   static final light = MoonModalColors(
+    textColor: MoonTypography.light.colors.bodyPrimary,
+    iconColor: MoonIconTheme.light.colors.primaryColor,
     backgroundColor: MoonColors.light.gohan,
     barrierColor: MoonColors.light.zeno,
   );
 
   static final dark = MoonModalColors(
+    textColor: MoonTypography.dark.colors.bodyPrimary,
+    iconColor: MoonIconTheme.dark.colors.primaryColor,
     backgroundColor: MoonColors.dark.gohan,
     barrierColor: MoonColors.dark.zeno,
   );
+
+  /// Modal text color.
+  final Color textColor;
+
+  /// Modal icon color.
+  final Color iconColor;
 
   /// Modal background color.
   final Color backgroundColor;
@@ -23,16 +35,22 @@ class MoonModalColors extends ThemeExtension<MoonModalColors> with Diagnosticabl
   final Color barrierColor;
 
   const MoonModalColors({
+    required this.textColor,
+    required this.iconColor,
     required this.backgroundColor,
     required this.barrierColor,
   });
 
   @override
   MoonModalColors copyWith({
+    Color? textColor,
+    Color? iconColor,
     Color? backgroundColor,
     Color? barrierColor,
   }) {
     return MoonModalColors(
+      textColor: textColor ?? this.textColor,
+      iconColor: iconColor ?? this.iconColor,
       backgroundColor: backgroundColor ?? this.backgroundColor,
       barrierColor: barrierColor ?? this.barrierColor,
     );
@@ -43,6 +61,8 @@ class MoonModalColors extends ThemeExtension<MoonModalColors> with Diagnosticabl
     if (other is! MoonModalColors) return this;
 
     return MoonModalColors(
+      textColor: colorPremulLerp(textColor, other.textColor, t)!,
+      iconColor: colorPremulLerp(iconColor, other.iconColor, t)!,
       backgroundColor: colorPremulLerp(backgroundColor, other.backgroundColor, t)!,
       barrierColor: colorPremulLerp(barrierColor, other.barrierColor, t)!,
     );
@@ -53,6 +73,8 @@ class MoonModalColors extends ThemeExtension<MoonModalColors> with Diagnosticabl
     super.debugFillProperties(properties);
     properties
       ..add(DiagnosticsProperty("type", "MoonModalColors"))
+      ..add(ColorProperty("textColor", textColor))
+      ..add(ColorProperty("iconColor", iconColor))
       ..add(ColorProperty("backgroundColor", backgroundColor))
       ..add(ColorProperty("barrierColor", barrierColor));
   }