+++ lib/src/theme/tag/tag_colors.dart
@@ -2,30 +2,48 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'package:moon_design/src/theme/colors.dart';
+import 'package:moon_design/src/theme/icons/icon_theme.dart';
+import 'package:moon_design/src/theme/typography/typography.dart';
 import 'package:moon_design/src/utils/color_premul_lerp.dart';
 
 @immutable
 class MoonTagColors extends ThemeExtension<MoonTagColors> with DiagnosticableTreeMixin {
   static final light = MoonTagColors(
+    textColor: MoonTypography.light.colors.bodyPrimary,
+    iconColor: MoonIconTheme.light.colors.primaryColor,
     backgroundColor: MoonColors.light.gohan,
   );
 
   static final dark = MoonTagColors(
+    textColor: MoonTypography.dark.colors.bodyPrimary,
+    iconColor: MoonIconTheme.dark.colors.primaryColor,
     backgroundColor: MoonColors.dark.gohan,
   );
+
+  /// Tag text color.
+  final Color textColor;
+
+  /// Tag icon color.
+  final Color iconColor;
 
   /// Tag background color.
   final Color backgroundColor;
 
   const MoonTagColors({
+    required this.textColor,
+    required this.iconColor,
     required this.backgroundColor,
   });
 
   @override
   MoonTagColors copyWith({
+    Color? textColor,
+    Color? iconColor,
     Color? backgroundColor,
   }) {
     return MoonTagColors(
+      textColor: textColor ?? this.textColor,
+      iconColor: iconColor ?? this.iconColor,
       backgroundColor: backgroundColor ?? this.backgroundColor,
     );
   }
@@ -35,6 +53,8 @@ class MoonTagColors extends ThemeExtension<MoonTagColors> with DiagnosticableTre
     if (other is! MoonTagColors) return this;
 
     return MoonTagColors(
+      textColor: colorPremulLerp(textColor, other.textColor, t)!,
+      iconColor: colorPremulLerp(iconColor, other.iconColor, t)!,
       backgroundColor: colorPremulLerp(backgroundColor, other.backgroundColor, t)!,
     );
   }
@@ -44,6 +64,8 @@ class MoonTagColors extends ThemeExtension<MoonTagColors> with DiagnosticableTre
     super.debugFillProperties(properties);
     properties
       ..add(DiagnosticsProperty("type", "MoonTagColors"))
+      ..add(ColorProperty("textColor", textColor))
+      ..add(ColorProperty("iconColor", iconColor))
       ..add(ColorProperty("backgroundColor", backgroundColor));
   }
 }